requirements.txt
asgiref==3.2.8
Django==3.0.7
Pillow==7.1.2
pkg-resources==0.0.0
pyserial==3.4
pytz==2020.1
sqlparse==0.3.1
